Code/3simulation/requirements.txt
torch>=1.13.0
open_clip_torch>=2.20.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0
networkx>=2.6.0
umap-learn>=0.5.0
hdbscan>=0.8.0
matplotlib>=3.5.0
tqdm>=4.62.0
